requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.7.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.1
redis==5.0.7
marshmallow==3.21.3
bcrypt==4.1.3
python-dotenv==1.0.1
Flask-Cors==4.0.0

# Testing
pytest==8.2.2
pytest-flask==1.3.0
Flask-Migrate==4.0.7
